Support maps in data classes
