+++ deep-print-reflection/src/main/java/com/bradyaiello/deepprint/Reflection.kt
@@ -48,7 +48,20 @@ fun Any?.deepPrintReflection(
                     )
                 },\n"
             )
-        } else {
+        } else if (propValue is Map<*,*>) {
+            builder.append(
+                "$initialIndent$indentIncrement$propName = ${
+                    propValue.deepPrintMapReflection(
+                        startingIndent = initialIndentLength + indentIncrementLength,
+                        indentSize = indentIncrementLength,
+                        standalone = false,
+                        constructor = "mutableMapOf",
+                    )
+                },\n"
+            )
+        }
+        
+        else {
             builder.append(
                 propValue.deepPrintReflection(
                     initialIndentLength = initialIndentLength + indentIncrementLength,
@@ -72,7 +85,7 @@ private fun Any.getPropertyValue(kParam: KParameter): Any? {
         .first { prop -> prop.name == kParam.name }.get(this)
 }
 
-fun <T : Any> KClass<T>.isPrimitive(): Boolean {
+internal fun <T : Any> KClass<T>.isPrimitive(): Boolean {
     return when (this) {
         Byte::class,
         Char::class,
@@ -100,137 +113,4 @@ fun Any.isPrimitive(): Boolean {
         is Double -> true
         else -> false
     }
-}
-
-fun <T> MutableList<T>.deepPrintMutableListReflection(
-    startingIndent: Int = 0,
-    indentSize: Int = 4,
-    standalone: Boolean = true,
-): String {
-    return this.deepPrintListReflection(
-        startingIndent = startingIndent,
-        indentSize = indentSize,
-        constructor = "mutableListOf",
-        standalone = standalone,
-    )
-}
-
-fun <T> List<T>.deepPrintListReflection(
-    startingIndent: Int = 0,
-    indentSize: Int = 4,
-    constructor: String = "listOf",
-    standalone: Boolean = true,
-): String {
-    val stringBuilder = StringBuilder()
-    val start = startingIndent.indent()
-    val prefix = if (standalone) start else " "
-    stringBuilder.append("${prefix}$constructor(\n")
-    this.forEach { value ->
-        value.deepPrintListItem(stringBuilder, startingIndent, indentSize)
-    }
-    stringBuilder.append("${start})")
-    return stringBuilder.toString()
-}
-
-fun <K, V> MutableMap<K, V>.deepPrintMutableMapReflection(
-    startingIndent: Int = 0,
-    indentSize: Int = 4,
-    constructor: String = "mutableMapOf",
-    standalone: Boolean = true,
-): String {
-    return deepPrintMapReflection(
-        startingIndent, 
-        indentSize, 
-        constructor, 
-        standalone,
-    )
-}
-
-
-fun <K, V> Map<K, V>.deepPrintMapReflection(
-    startingIndent: Int = 0,
-    indentSize: Int = 4,
-    constructor: String = "mapOf",
-    standalone: Boolean = true,
-): String {
-    val stringBuilder = StringBuilder()
-    val start = startingIndent.indent()
-    val prefix = if (standalone) start else " "
-    stringBuilder.append("${prefix}$constructor(\n")
-
-    this.forEach { entry ->
-        entry.deepPrintMapEntryReflection(
-            stringBuilder = stringBuilder,
-            startingIndent = startingIndent,
-            indentSize = indentSize,
-        )
-    }
-    stringBuilder.append("${start})")
-    return stringBuilder.toString()
-}
-
-private fun <Any> Any?.deepPrintListItem(
-    stringBuilder: StringBuilder,
-    startingIndent: Int,
-    indentSize: Int
-) {
-    val totalIndent = startingIndent.indent() + indentSize.indent()
-
-    if (this == null) {
-        stringBuilder.append("${totalIndent}null,\n")
-    } else if (this!!::class.isPrimitive()) {
-        stringBuilder.append("${totalIndent}${deepPrintPrimitive(this)},\n")
-    } else {
-        stringBuilder.append(
-            this.deepPrintReflection(
-                initialIndentLength = startingIndent + indentSize,
-                indentIncrementLength = indentSize,
-            ) + "\n"
-        )
-    }
-}
-
-private fun <Any> Any?.deepPrintMapKeyOrValue(
-    stringBuilder: StringBuilder,
-    startingIndent: Int,
-    indentSize: Int
-) {
-    val totalIndent = startingIndent.indent() + indentSize.indent()
-
-    if (this == null) {
-        stringBuilder.append("${totalIndent}null")
-    } else if (this!!::class.isPrimitive()) {
-        stringBuilder.append("${totalIndent}${deepPrintPrimitive(this)}")
-    } else {
-        stringBuilder.append(
-            this.deepPrintReflection(
-                initialIndentLength = startingIndent + indentSize,
-                indentIncrementLength = indentSize,
-            )
-        )
-    }
-}
-
-private fun <K, V> Map.Entry<K, V?>.deepPrintMapEntryReflection(
-    stringBuilder: StringBuilder,
-    startingIndent: Int,
-    indentSize: Int,
-) {
-    key.deepPrintMapKeyOrValue(
-        stringBuilder = stringBuilder, 
-        startingIndent = startingIndent, 
-        indentSize = indentSize
-    )
-    val singleLine = key!!.isPrimitive() && ((value == null) || value!!.isPrimitive())
-    val toStr = if (singleLine) " to " else " to\n"
-    stringBuilder.append(toStr)
-    val newStartingIndent = if (singleLine) 0 else startingIndent + indentSize
-    val newIndentSize = if (singleLine) 0 else indentSize
-    value.deepPrintMapKeyOrValue(
-        stringBuilder = stringBuilder,
-        startingIndent = newStartingIndent,
-        indentSize = newIndentSize
-    )
-    val suffix = if (singleLine) ",\n" else "\n"
-    stringBuilder.append(suffix)
 }
